refactor: separated widgets into their own files and fixed file structure
