+++ mediguide/lib/presentation/screens/chat_screen/widgets/bot_chat_bubble.dart
