+++ mediguide/lib/presentation/screens/chat_screen/widgets/chat_input_section.dart
