+++ mediguide/lib/presentation/screens/chat_screen/widgets/navigation_drawer.dart
@@ -1,60 +1,10 @@
+
 import 'package:flutter/material.dart';
-// import 'package:provider/provider.dart';
-
-import 'package:mediguide/enums/menu_actions.dart';
-import 'package:mediguide/presentation/screens/sign_in_screen.dart';
-import 'package:mediguide/utils/theme_constants.dart';
-// import 'package:mediguide/utils/theme_manager.dart';
-import 'package:mediguide/utils/theme_utils.dart';
-import 'package:mediguide/utils/custom_icons.dart';
+import 'package:mediguide/presentation/screens/auth_screen/sign_in_screen.dart';
+import 'package:mediguide/presentation/screens/chat_screen/widgets/custom_chat_history_tile.dart';
 import 'package:mediguide/presentation/widgets/settings.dart';
-import 'package:mediguide/presentation/widgets/user_chat_bubble.dart';
-import 'package:mediguide/presentation/widgets/bot_chat_bubble.dart';
-import 'package:mediguide/presentation/widgets/chat_input_section.dart';
-
-
-class ChatScreen extends StatelessWidget {
-  const ChatScreen({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    // final themeManager = Provider.of<ThemeManager>(context);
-
-    return Scaffold(
-      appBar: const PreferredSize(
-        preferredSize: Size.fromHeight(70),
-        child: CustomAppBar(),
-      ),
-      body: Column(
-        children: [
-          Expanded(
-            child: ListView(
-              padding: const EdgeInsets.all(15.0),
-              children: const <Widget>[
-                BotChatBubble(
-                    message:
-                        "Hi! I am MediGuide. How can I assist you today? If you're not feeling your best, just let me know your symptoms, and I'll do my best to provide you with some insights. Remember, I'm here to help, but for accurate medical advice, always consult a healthcare professional."),
-                UserChatBubble(
-                    message: "shivering, chills, joint pain, headache"),
-                BotChatBubble(
-                    message:
-                        "Based on your symptoms, you might have gastroenteritis. Gastroenteritis is a short-term illness triggered by the infection and inflammation of the digestive system."),
-                UserChatBubble(
-                    message: "shivering, chills, joint pain, headache"),
-                BotChatBubble(
-                    message:
-                        "Based on your symptoms, you might have gastroenteritis. Gastroenteritis is a short-term illness triggered by the infection and inflammation of the digestive system.")
-              ],
-            ),
-          ),
-          const ChatInputSection(),
-        ],
-      ),
-      drawer: const NavigationDrawer(),
-    );
-  }
-}
-
+import 'package:mediguide/utils/theme_constants.dart';
+import 'package:mediguide/utils/theme_utils.dart';
 
 class NavigationDrawer extends StatelessWidget {
   const NavigationDrawer({Key? key}) : super(key: key);
@@ -274,152 +224,6 @@ class NavigationDrawer extends StatelessWidget {
           )
         ],
       ),
-    );
-  }
-}
-
-
-class CustomChatHistoryTile extends StatelessWidget {
-  final String title;
-  final VoidCallback onTap;
-  final VoidCallback onDelete;
-
-  const CustomChatHistoryTile({
-    super.key,
-    required this.title,
-    required this.onTap,
-    required this.onDelete,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    final ThemeData currentTheme = ThemeUtils.getTheme(context);
-
-    return Padding(
-      padding: const EdgeInsets.symmetric(horizontal: 5),
-      child: Row(
-        mainAxisAlignment: MainAxisAlignment.spaceBetween,
-        children: [
-          Expanded(
-            child: TextButton(
-              onPressed: () {
-                // Your onPressed logic here
-              },
-              style: TextButton.styleFrom(
-                backgroundColor: Colors.transparent,
-                elevation: 0,
-              ),
-              child: Row(
-                children: [
-                  Icon(Icons.message_outlined, size: 14, color: ThemeUtils.getForeground(currentTheme),),
-                  const SizedBox(width: 10.0), // Adjust spacing as needed
-                  Expanded(
-                    child: Text(
-                      title,
-                      overflow: TextOverflow.ellipsis,
-                      maxLines: 1,
-                      style: TextStyle(fontFamily: 'Poppins', fontSize: 12, color: ThemeUtils.getForeground(currentTheme)),
-                    ),
-                  ),
-                ],
-              ),
-            ),
-          ),
-          IconButton(
-            onPressed: () {
-              // delete button onPressed logic here
-            },
-            icon: const Icon(Icons.delete_outline_outlined, size: 14, color: Color(0xFFFF5959)),
-          )
-        ],
-      ),
-    );
-  }
-}
-
-class CustomAppBar extends StatelessWidget {
-  const CustomAppBar({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    final ThemeData currentTheme = ThemeUtils.getTheme(context);
-
-    return AppBar(
-      leadingWidth: 75,
-      toolbarHeight: 70,
-      backgroundColor: currentTheme.appBarTheme.backgroundColor,
-      leading: IconButton(
-        color: ThemeUtils.getForeground(currentTheme),
-        icon: Container(
-          padding: const EdgeInsets.all(10),
-          decoration: BoxDecoration(
-            color: ThemeUtils.getIconBackground(currentTheme),
-            borderRadius: const BorderRadius.all(Radius.circular(8.0)),
-          ),
-          child: const Icon(CustomIcons.ham, size: 20, color: accentColor),
-        ),
-        onPressed: () {
-          Scaffold.of(context).openDrawer();
-        },
-      ),
-      title: Center(
-        child: Text(
-          'New Chat',
-          style: TextStyle(
-            color: ThemeUtils.getForeground(currentTheme),
-            fontFamily: 'Poppins',
-            fontWeight: FontWeight.w400,
-            fontSize: 14,
-          ),
-        ),
-      ),
-      actions: const [AppBarPopupMenu()],
-    );
-  }
-}
-
-class AppBarPopupMenu extends StatelessWidget {
-  const AppBarPopupMenu({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    final ThemeData currentTheme = ThemeUtils.getTheme(context);
-
-    return Container(
-      width: 75,
-      padding: const EdgeInsets.only(left: 16.0),
-      child: PopupMenuButton<MenuAction>(
-          color: currentTheme.appBarTheme.backgroundColor,
-          icon: const Icon(Icons.more_vert, color: Color(0xFFAFAFAF)),
-          onSelected: (value) {
-            if (value == MenuAction.deleteChat) {
-              // perform action to delete chat
-            }
-            if (value == MenuAction.newChat) {
-              // perform action to create new chat
-            }
-            if (value == MenuAction.settings) {
-              showSettingsDialog(context);
-            }
-          },
-          offset: const Offset(0, 80),
-          itemBuilder: (context) => [
-                const PopupMenuItem(
-                  value: MenuAction.deleteChat,
-                  padding: EdgeInsets.symmetric(horizontal: 30.0),
-                  child: Text('Delete Chat', style: TextStyle(fontSize: 12)),
-                ),
-                const PopupMenuItem(
-                  value: MenuAction.newChat,
-                  padding: EdgeInsets.symmetric(horizontal: 30.0),
-                  child: Text('New Chat', style: TextStyle(fontSize: 12)),
-                ),
-                const PopupMenuItem(
-                  value: MenuAction.settings,
-                  padding: EdgeInsets.symmetric(horizontal: 30.0),
-                  child: Text('Settings', style: TextStyle(fontSize: 12)),
-                ),
-              ]),
     );
   }
 }
