+++ mediguide/lib/presentation/screens/chat_screen/widgets/user_chat_bubble.dart
